requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==4.2.16
django-allauth==64.2.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
idna==3.8
jwt==1.3.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
